Add commit-bound plugin verification
