+++ AGENTS.md
@@ -128,7 +128,9 @@ New automated submissions use one public GitHub repository per plugin, with `man
 
 The Automated Security Baseline statically scans the exact validated commit without executing plugin code. `SECURITY_BASELINE.md` is the public policy for all contributors; keep implementation, reports, tests, and contributor documentation aligned with it. Preserve deterministic outcomes (`passed`, `review-required`, and `needs-fixes`), evidence templates, exact-SHA binding, and the required disclaimer. Do not use AI to determine outcomes, enforcement, labels, or approval. In selective enforcement mode, only documented critical findings use `security-needs-fixes` and block approval; capabilities and non-enforced findings use `security-review-required` and remain eligible for explicit maintainer approval. Scan failures remain fail-closed. Treat executable binaries as review capabilities, narrowly scoped root-owned helpers as review cases rather than automatic blockers, and execution sourced only from the submitted repository as self-installation requiring review rather than as external mutable-code rejection.
 
-The registry stores automated baseline facts, including finding and capability IDs, but no maintainer identity, review timestamp, review flag, or other copied approval provenance. The baseline is a listing-time check. Scheduled catalog refreshes still inspect upstream branch HEAD for compatibility and do not refresh the stored baseline. Do not describe that stored result as applying to later upstream commits.
+Security module boundaries are deliberate. `security-baseline-policy.mjs` is the single owner of policy version, marker protocol, finding/capability catalogs, outcome, enforcement, and label disposition. `security-baseline-record.mjs` is the only stored-record and marker converter. `security-github-snapshot.mjs` encapsulates bounded GitHub transport, `security-baseline-scope.mjs` owns deterministic file selection, `security-baseline-analysis.mjs` owns static findings and capabilities, and `security-baseline-scanner.mjs` only orchestrates scope plus analysis. `security-baseline.mjs` is the compatible CLI/facade. `verification-subject.mjs` resolves exact source-wide scan plans by configured plugin ID, and `catalog-verification.mjs` is the only catalog status projection. Manifest paths from mutable catalog refreshes are hints only. Catalog projection must require the exact registry plugin set for a source and fail closed on stale extra entries. Domain modules must not import filesystem APIs, `sharp`, workflows, or `build-catalog.mjs`. Approval and verification must persist identical canonical baseline records.
+
+The registry stores automated baseline facts, including record schema, repository, affected plugin IDs, finding IDs, and capability IDs, but no maintainer identity, review timestamp, review flag, or other copied approval provenance. The baseline is a listing-time check. Scheduled catalog refreshes still inspect upstream branch HEAD for compatibility and do not refresh the stored baseline. Do not describe that stored result as applying to later upstream commits. Derive `Verified` only from a current-version `passed` baseline with empty finding and capability sets whose commit equals `listingValidatedCommit`; otherwise derive `Unverified`. Verification is automatic, has no manual override, and describes only the exact listed commit, never trust or later upstream code.
 
 Root community plugins default to the generated standard Omarchy install command. Use a manual installation override only when that command cannot produce a functioning plugin, such as when a required native binary is absent or a mandatory build must run before enablement. Optional feature setup, credentials, API configuration, data sources, bar placement, and layout customization do not make an otherwise functioning plugin manual. A curated `plugins.<id>.installation` override in `registry.json` must contain exactly `mode: "manual"` and a non-empty user-facing `note`. Manual overrides must not publish arbitrary installer commands, apply only to root plugin repositories, and must keep installation disabled and the curated note intact across passed, failed, and unreachable upstream checks. During automated listing review, apply `manual-setup` before `approved-for-listing`; approval maps that maintainer decision to the registry override.
