+++ SUBMISSION.md
@@ -133,3 +133,7 @@ The marketplace bot keeps one validation comment and one automated-security-base
 Approval and publication failures use a separate status comment with the failed phase, a safe error summary, the required action, and a workflow link. Baseline approval is bound to the exact listing commit. The registry records the automated result and exact commit but does not copy maintainer identities, review timestamps, or review flags into permanent Git history. Later scheduled catalog refreshes continue to inspect the repository's branch head for compatibility and do not rerun the listing-time security baseline; the stored baseline therefore describes only the commit approved for initial listing. When approval fails, the `approved-for-listing` label is removed automatically. Fix the reported problem and ask a maintainer to reapply the label after validation passes; rerunning the old failed workflow does not restore the label. If registration succeeded but deployment or issue finalization failed, do not resubmit the plugin or reapply the approval label—a maintainer must retry or complete the reported publication phase.
 
 If no automated validation comment appears, edit the existing issue and verify that its title starts with `[Plugin]:`, all six headings remain in their original order, the category matches exactly, and all five checklist items are checked. Editing the issue runs submission detection again.
+
+## Verify an existing listing
+
+Use the **Request plugin verification** issue form instead of opening another submission. The request must identify the existing plugin ID, repository, and full `listingValidatedCommit` exactly. The automated workflow reruns the static baseline only for that recorded commit and publishes `Verified` only after a complete `passed` result with no findings or review capabilities. Other results remain `Unverified`; this is not a claim that the plugin is malicious. Read [Automated Plugin Verification](VERIFICATION.md) for the complete status definition, process, and limits.
