+++ scripts/security-baseline-limits.mjs
@@ -1,0 +1,4 @@
+export const securityFileByteLimit = 512 * 1024;
+export const securitySnapshotByteLimit = 8 * 1024 * 1024;
+export const securitySnapshotFileLimit = 1000;
+export const securityBinaryProbeByteLimit = 4096;
