+++ site/assets/js/develop.js
@@ -2,7 +2,7 @@ import {
   setupCopyButtons,
   setupSectionNavigation,
   setupThemeToggle,
-} from "./shared.js?v=20260816-11";
+} from "./shared.js?v=20260816-12";
 
 setupThemeToggle();
 setupCopyButtons();
