+++ test/automation.test.js
@@ -75,6 +75,7 @@ import {
   appendCatalogViewState,
   catalogViewControls,
   findCopyLabel,
+  pluginVerificationState,
   readCatalogViewState,
   showCopiedState,
   writeClipboard,
@@ -456,6 +457,20 @@ test("search keeps the raw query until a completion is explicitly accepted", () 
   }), "accept-active-completion");
   assert.equal(searchKeyAction({ ...defaults, key: "ArrowDown" }), "next-completion");
   assert.equal(searchKeyAction({ ...defaults, key: "ArrowUp" }), "previous-completion");
+});
+
+test("plugin verification display copy is minimal and distinguishes third-party statuses", () => {
+  assert.deepEqual(pluginVerificationState({ verificationStatus: "verified" }), {
+    status: "verified",
+    label: "Verified",
+    explanation: "Automated checks passed for the listed commit. This is not a security audit.",
+  });
+  assert.deepEqual(pluginVerificationState({ verificationStatus: "unverified" }), {
+    status: "unverified",
+    label: "Unverified",
+    explanation: "No passing automated baseline is recorded for the listed commit. This does not mean the plugin is malicious.",
+  });
+  assert.equal(pluginVerificationState({ builtIn: true }), null);
 });
 
 test("copy feedback targets the visible label instead of a decorative icon", () => {
@@ -933,7 +948,18 @@ test("entry modules and their shared dependency use one cache key", async () => 
   assert.doesNotMatch(files.app, /plugin-preview-(?:bar|meta)/);
   assert.match(styles, /\.plugin-card-body \.plugin-description \{[\s\S]*max-height: 21px;[\s\S]*text-overflow: clip; white-space: nowrap;[\s\S]*mask-image: linear-gradient/);
   assert.match(styles, /\.plugin-card-body \.plugin-description \{[\s\S]*margin: 4px 0 9px;/);
+  assert.match(styles, /\.card-status-line \{[\s\S]*justify-content: space-between;/);
   assert.match(styles, /\.card-activity-state \{[\s\S]*border-left: 2px solid currentColor;[\s\S]*font-size: 10px;/);
+  assert.match(styles, /\.card-verification-trigger \{[\s\S]*min-width: 24px; height: 24px;[\s\S]*text-transform: none;/);
+  assert.match(styles, /\.card-verification-marker \{[\s\S]*height: 21px;[\s\S]*border-right: 2px solid currentColor;/);
+  assert.match(styles, /\.card-verification-tooltip \{[\s\S]*right: 0;[\s\S]*bottom: calc\(100% \+ 7px\);/);
+  assert.match(styles, /\.card-verification:hover:not\(\.is-dismissed\) \.card-verification-tooltip,[\s\S]*\.card-verification:focus-within:not\(\.is-dismissed\) \.card-verification-tooltip,[\s\S]*\.card-verification\.is-open \.card-verification-tooltip/);
+  assert.match(files.app, /class="card-status-line">\$\{activityState\}\$\{verificationState\}/);
+  assert.match(files.app, /data-verification-tooltip aria-expanded="false" aria-label=/);
+  assert.match(files.app, /button\.addEventListener\("click", \(event\) => \{[\s\S]*classList\.toggle\("is-open", expanded\)/);
+  assert.match(files.app, /event\.key !== "Escape"/);
+  assert.match(files.app, /matches\?\.\("\[data-verification-tooltip\]"\)[\s\S]*control = "verification"/);
+  assert.match(files.app, /verification: "\[data-verification-tooltip\]"/);
   assert.doesNotMatch(styles, /\.plugin-title-line \.new-badge/);
   assert.match(styles, /\.card-social \{[\s\S]*top: 13px;[\s\S]*height: 25px;/);
   assert.match(styles, /\.card-stars \{[\s\S]*width: 54px;[\s\S]*height: 25px;/);
@@ -1001,7 +1027,7 @@ test("entry modules and their shared dependency use one cache key", async () => 
   assert.match(styles, /\.market-plugin-grid \.plugin-card, \.recent-grid \.plugin-card \{[\s\S]*display: flex;[\s\S]*flex-direction: column; gap: 0;/);
   assert.match(styles, /\.plugin-preview \{[\s\S]*height: 175px; min-height: 0;[\s\S]*flex: 0 0 175px;/);
   assert.match(styles, /\.plugin-card-body \{[\s\S]*display: flex;[\s\S]*min-width: 0;[\s\S]*flex: 1; flex-direction: column;/);
-  assert.match(files.app, /<div class="plugin-card-content">[\s\S]*class="plugin-title-line"[\s\S]*\$\{authorLine\}[\s\S]*class="plugin-description"[\s\S]*\$\{activityState\}/);
+  assert.match(files.app, /<div class="plugin-card-content">[\s\S]*class="plugin-title-line"[\s\S]*\$\{authorLine\}[\s\S]*class="plugin-description"[\s\S]*\$\{cardStates\}/);
   assert.match(styles, /\.plugin-card-bottom \{[\s\S]*margin-top: auto;/);
   assert.match(styles, /\.plugin-author button \{[\s\S]*z-index: 3/);
   assert.match(styles, /\.plugin-author button \{[\s\S]*min-height: 24px/);
@@ -1298,17 +1324,18 @@ test("automation deploys refreshed catalogs and uses listing-specific approval",
   assert.match(validate, /--metadata=validation-metadata\.json/);
   assert.match(validate, /--json=security-baseline\.json/);
   assert.match(validate, /passed\|review-required\|needs-fixes/);
-  assert.match(validate, /marketplace-security-baseline:v\[123\]/);
-  assert.match(validate, /marketplace-security-baseline-error:v\[123\]/);
+  assert.match(validate, /marketplace-security-baseline:v\[0-9\]\+/);
+  assert.match(validate, /marketplace-security-baseline-error:v\[0-9\]\+/);
   assert.match(validate, /--add-label security-needs-fixes/);
   assert.match(validate, /--add-label security-review-required/);
-  assert.match(validate, /blocks_approval="\$\(jq -r '\.blocksApproval' security-baseline\.json\)"/);
+  assert.match(validate, /disposition="\$\(jq -r '\.disposition' security-baseline\.json\)"/);
+  assert.match(validate, /clear\|review-required\|needs-fixes/);
+  assert.match(validate, /BASELINE_DISPOSITION: \$\{\{ needs\.validate\.outputs\.baseline_disposition \}\}/);
   assert.match(
     validate,
-    /needs-fixes\)\s+[\s\S]*?BASELINE_BLOCKS_APPROVAL[\s\S]*?--add-label security-needs-fixes[\s\S]*?--add-label security-review-required/,
+    /case "\$BASELINE_DISPOSITION"[\s\S]*clear\)[\s\S]*review-required\)[\s\S]*needs-fixes\)/,
   );
-  assert.match(validate, /BASELINE_RESULT: \$\{\{ needs\.validate\.outputs\.baseline_result \}\}/);
-  assert.match(validate, /BASELINE_BLOCKS_APPROVAL: \$\{\{ needs\.validate\.outputs\.baseline_blocks_approval \}\}/);
+  assert.doesNotMatch(validate, /BASELINE_BLOCKS_APPROVAL|baseline_blocks_approval/);
   assert.match(validate, /name: Clear stale approval state after workflow failure/);
   assert.match(validate, /labels\/\$\{label\}/);
   assert.match(validate, /remove_label approved-for-listing/);
@@ -1920,8 +1947,8 @@ test("approved submissions become registry sources without duplicates", () => {
       tags: ["hyprland", "workspaces"],
     },
     manifests: [
-      { id: "example.overview", name: "Overview" },
-      { id: "example.switcher", name: "Switcher" },
+      { id: "example.overview", name: "Overview", path: "overview/manifest.json" },
+      { id: "example.switcher", name: "Switcher", path: "switcher/manifest.json" },
     ],
     addedAt: "2026-07-28",
     listedAt: "2026-07-28T11:17:52.000Z",
@@ -1942,24 +1969,30 @@ test("approved submissions become registry sources without duplicates", () => {
       "example.overview": {
         category: "Desktop",
         tags: ["hyprland", "workspaces"],
+        manifestPath: "overview/manifest.json",
       },
       "example.switcher": {
         category: "Desktop",
         tags: ["hyprland", "workspaces"],
+        manifestPath: "switcher/manifest.json",
       },
     },
   });
   const baselineRecord = createApprovedSecurityBaseline({
     baselineVersion: "3",
+    repository: "example/plugin",
     commitSha: "c".repeat(40),
     checkedAt: "2026-07-28T11:00:00.000Z",
     outcome: "review-required",
     enforcementMode: "selective",
     findings: [],
     capabilities: ["service-management"],
-  });
+  }, { pluginIds: ["example.plugin"] });
   assert.deepEqual(baselineRecord, {
+    schemaVersion: 1,
     version: "3",
+    repository: "example/plugin",
+    pluginIds: ["example.plugin"],
     commit: "c".repeat(40),
     checkedAt: "2026-07-28T11:00:00.000Z",
     outcome: "review-required",
@@ -1989,13 +2022,14 @@ test("approved submissions become registry sources without duplicates", () => {
   });
   const selectiveReviewRecord = createApprovedSecurityBaseline({
     baselineVersion: "3",
+    repository: "example/plugin",
     commitSha: "d".repeat(40),
     checkedAt: "2026-07-28T11:00:00.000Z",
     outcome: "needs-fixes",
     enforcementMode: "selective",
     findings: ["remote-git-execution-unpinned"],
     capabilities: ["remote-build"],
-  });
+  }, { pluginIds: ["example.plugin"] });
   assert.equal(selectiveReviewRecord.outcome, "needs-fixes");
   assert.deepEqual(selectiveReviewRecord.findings, ["remote-git-execution-unpinned"]);
   assert.equal(Object.hasOwn(selectiveReviewRecord, "reviewedBy"), false);
@@ -2003,13 +2037,14 @@ test("approved submissions become registry sources without duplicates", () => {
   assert.throws(
     () => createApprovedSecurityBaseline({
       baselineVersion: "3",
+      repository: "example/plugin",
       commitSha: "e".repeat(40),
       checkedAt: "2026-07-28T11:00:00.000Z",
       outcome: "passed",
       enforcementMode: "selective",
       findings: ["curl-pipe-shell"],
       capabilities: [],
-    }),
+    }, { pluginIds: ["example.plugin"] }),
     (error) => error.code === "approval-security-baseline-invalid",
   );
   assert.equal(parseManualSetupApproval("true"), true);
